requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
bcrypt==4.3.0
binaryornot==0.4.4
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
et_xmlfile==2.0.0
fastapi==0.101.1
flet==0.19.0
flet-core==0.19.0
flet-fastapi==0.19.0
flet-runtime==0.19.0
greenlet==3.2.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.3.1
oauthlib==3.3.1
openpyxl==3.1.5
packaging==23.2
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyodbc==5.2.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.4
rich==14.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.27.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==3.0.0
websocket-client==1.8.0
uvicorn==0.29.0
websockets==11.0.3
python-dotenv==1.1.1
